Patches provided by Ake Sandgren and Robert Granat --------------------------------------------------- The set of patches for scalapack 1.7.4 does two things. 1 - reduce the usage of uninitialized variables 2 - fix a couple of incorrect calls to blacs (bad LDA)
The gehdrv patch is just the complete patch related to
https://icl.cs.utk.edu/lapack-forum/viewtopic.php?p=1153#1153

pxsepinfo doesnt initialize THRESH when INFO != 0.

I'm not sure if the pxlahrd patch is the best. Maybe something should be
done in pxlarfg instead since alpha isn't set in all cases there, like
myrow != ixrow for row distribution and likewise for column
distribution.

pxlasmsub destroys irow1/icol1 in the "find some norm of the local H"
part.

pxrot used incorrect LDA values for buff in several places, not sure if
the intention was to have buff Mx1 or 1xM but it shouldn't really matter
should it?

PBLAS/pxscal must not test ALPHA unless it is really going to be used
since scalapack routines sometimes call pxscal with ALPHA uninitialized
when myrow != Xrow/mycol != Xcol.

pxstein must initialize ONENRM since it isn't always initialized in the
"IF( NBLK.EQ.IBLOCK( NEXT-1 ) .AND. NBLK.NE.OLNBLK ) THEN" case before
being used in the "IF( TMPFAC.GT.ODM18 ) THEN" case. Maybe setting to
ZERO is wrong but its not worse then the original code.

pxtrevc and pxevcdriver are just incorrect LDA param to blacs routines.

All these was found with pathscale compiler with -trapuv -O0 -g which
initialized everything to NaN and turns FPE traps on.

--------------------------------------------------------------------------

The lahqr patch and a fix to lasorte needed by lahqr which
used to get IERR != 0 back from lasorte.

The T2 = T1*V2 and T3 = T1*V3 moves are needed due to uninitialized
data.
The 2 changed IF-statements where brought about to make getting and
sending SMALLA consistent.
The ISTOP change at the bottom is a copy of the corresponding statement
at the top of the loop.

The init of VCOPY and SMALLA are neccesary.

lasorte couldn't handle a situation where the top S(1,1) eigenvalue was
real.

This set of patches have been tested as can be seen on
https://icl.cs.utk.edu/lapack-forum/viewtopic.php?p=1196#1196
so i'm fairly certain that they work correctly.