Modifications following the bug report ot Yana Gurieva on the Forum (topic: 2406)
"Rev.82 of Scalapack pslascl.f and pclascl.f still have a possibility for incorrect work: 
these sources are intended for single precision computations but line 192 inside them does
contain a call to DISNAN function. Shouldn't it be SISNAN instead?

As for PBtools.h, Scalapack's rev.69 changes Mptr inside, but in its definition in line 130
after the symbol '\' there should be a carriage return. The current line produces a compiler
error. Moreover, the definition of (off_t) used here exists only for Linux. For Windows, it
is undefined and the package cannot be compiled and linked."


Change off_t by long long, needs to test it under Windows.