Allow compilation in ILP64 mode
The code now uses preprocessor macro `Int` instead of the `int` type
to allow for flexibility in compilation of the library. `Int` is the
internal integer type used for most operations. Also, it is the type
used (expected) by the library interface. By default `Int` is `int`.
However, it is possible to redefine it on the command line to a
larger integer type `long` if desired.

Any change to `Int` needs to be accompanied with the appropriate
promotion of the Fortran integers (`-fdefault-integer-8`, `-i8`
etc.).

When compiling ILP64 version of ScaLAPACK, the following needs to be
satisfied:
  1. Use ILP64 BLAS (e.g. OpenBLAS with `-DINTERFACE64=ON`).
  2. Promote integer types (`-DInt=long`, `-fdefault-integer-8`).
  3. Change hardcoded `INTSZ` and `INTGSZ` in the test sources
     from 4 to 8.