enh: no memory leaks for blacs_exit and allow re-call
This should enable ScaLAPACK/BLACS to be exited and restarted
at will without memory leaks or problems.

When called with blacs_exit(true), the comm_world will be
retained and MPI_Finalized won't be called.
Only when blacs_exit(false) is called will the BI_COMM_WORLD
be de-allocated. So one should expect a memory leak of
an integer if the hosting application calls MPI_Finalize.