set project version and shared library soname
Set the ScaLAPACK project version in a single place, at the project
command in CMakeLists.txt.

This defines SCALAPACK_VERSION, from which cmake derives
SCALAPACK_VERSION_MAJOR, SCALAPACK_VERSION_MINOR, SCALAPACK_VERSION_PATCH

Use SCALAPACK_VERSION as the shared library VERSION (defining the
shared library file as libscalapack.so.2.2.1)

and use SCALAPACK_VERSION_MAJOR.SCALAPACK_VERSION_MINOR as the shared
library SOVERSION (defining the soname as libscalapack.so.2.2)

That is, assume that the ABI updates with each MINOR release and
reflect that in the soname.

An API update (and therefore an ABI update) would be expected with any
MAJOR release.

No ABI update should occur with a PATCH release.

Closes: #46